extract shouldInitialize to Source
